feat(Impl): Add CI/CD pipeline, update readme and use correct type to represent index dimension (uint)

+++ src/IndexTypes.cs
@@ -12,7 +12,7 @@ public readonly record struct IndexName(string Value)
 public record IndexDetails
 {
     public required string Name { get; init; }
-    public required long Dimension { get; init; }
+    public required uint Dimension { get; init; }
     public required Metric Metric { get; init; }
     public long? Pods { get; init; }
     [JsonPropertyName("pod_type")]
